Unity to Arduino Serial Communication
Multithreading allows operations to run in parallel and maintain the refresh rate of each script (smooth haptic rendering requires an update rate of 1kHz).  The force calculation is performed on a physics engine and a lock is used to prevent both scripts accessing the resource at the same time which could cause incorrect data transfer.